debian: enhance clean target dependencies
When building a package with pbuilder, the clean target is run outside
of the chroot. That causes unmet build dependencies with
python-sphinxdoc and python-mock.  They are only needed for the binary
package but Debian does not have a way to set different dependencies to
build the source package.

Use a standard clean target so pbuilder does not complain about unmet
build dependencies
Add dh-python to build dependencies. On Jessie, it is a dependencies of
the python3 package, but might not always be installed nonetheless.

Signed-off-by: Alexandros Kosiaris <akosiaris@gmail.com>