Decode the command output
subprocess.Popen returns a bytes string, both on Python 2 and Python 3.

On Python 2 everything works because Python 2 automatically tries to
decode/encode strings as needed. (often badly)

However on Python 3 the rest of the code, passed to
subprocess.check_call, expects a unicode string.

Always decoding the output of the Popen command makes this work both on
Python 2 and 3.