Use the right StringIO/BytesIO
The tests try to mock sys.stderr so that they can verify what was
printed there.

They use StringIO.StringIO to do that.

However, that class does not exist on Python 3, it's instead
io.StringIO.

Unfortunately, when optparse can't parse the arguments it will print
str(error_message) to sys.stderr, both on Python 2 and 3.

That means optparse ends up sending to sys.stderr a bytes string on
Python 2, but a unicode string on Python 3.

To counter that, we must mock sys.stderr with io.BytesIO on Python 2,
and io.StringIO on Python 3.