bootstrap python3 venv with pip and setuptools when --system-site-package is used
Due to Python bug https://bugs.python.org/issue24875 the `venv` module does
not install pip and setuptools when `--system-site-package` is used.

This has been fixed upstream[1] but has not been backported by distribution maintainers.

The accepted fix is to create the virtual environment first without `--system-site-package`
allowing pip and setuptools to be installed and then reconfiguring the environment with
`--system-site-package` to enable this option. This is also the recommend fix found by users
on Stack Overflow[2].

[1] https://github.com/python/cpython/blob/e42b705188271da108de42b55d9344642170aa2b/Lib/venv/__init__.py#L61
[2] https://stackoverflow.com/a/40989759/388127

Signed-off-by: Lucas Magasweran <lucas.magasweran@ieee.org>